+++ tests/testthat/test-textmodel_newsmap.R
@@ -1,9 +1,9 @@
 require(quanteda)
 
-test_that("test English dictionary and prediction work correctly", {
-    text_en <- c("This is an article about Ireland.")
+test_that("English dictionary and prediction work correctly", {
 
-    toks_en <- tokens(text_en)
+    txt_en <- c("This is an article about Ireland.")
+    toks_en <- tokens(txt_en)
     label_toks_en <- tokens_lookup(toks_en, data_dictionary_newsmap_en, levels = 3)
     label_dfm_en <- dfm(label_toks_en)
 
@@ -17,10 +17,10 @@ test_that("test English dictionary and prediction work correctly", {
 })
 
 
-test_that("test German dictionary and prediction work correctly", {
-    text_de <- c("Ein Artikel über Irland.")
+test_that("German dictionary and prediction work correctly", {
 
-    toks_de <- tokens(text_de)
+    txt_de <- c("Ein Artikel über Irland.")
+    toks_de <- tokens(txt_de)
     label_toks_de <- tokens_lookup(toks_de, data_dictionary_newsmap_de, levels = 3)
     label_dfm_de <- dfm(label_toks_de)
 
@@ -34,10 +34,57 @@ test_that("test German dictionary and prediction work correctly", {
     )
 })
 
-test_that("test Japanese dictionary and prediction work correctly", {
-    text_ja <- c("アイルランドに関するテキスト.")
+test_that("test French dictionary and prediction work correctly", {
 
-    toks_ja <- tokens(text_ja)
+    txt_fr <- c("Ceci est un article sur l'Irlande.")
+    toks_fr <- tokens(txt_fr)
+    toks_fr <- tokens_split(toks_fr, "'")
+    label_toks_fr <- tokens_lookup(toks_fr, data_dictionary_newsmap_fr, levels = 3)
+    label_dfm_fr <- dfm(label_toks_fr)
+
+    feat_dfm_fr <- dfm(toks_fr, tolower = FALSE) %>%
+        dfm_select('^[A-Z][A-Za-z1-2]+', selection = "keep",
+                   valuetype = 'regex', case_insensitive = FALSE)
+
+    expect_equal(
+        as.character(predict(textmodel_newsmap(feat_dfm_fr, label_dfm_fr))),
+        "ie"
+    )
+})
+
+test_that("Hebrew dictionary and prediction work correctly", {
+
+    skip_on_travis()
+    txt_he <- c("טקסט על אירלנד.")
+    toks_he <- tokens(txt_he)
+    label_toks_he <- tokens_lookup(toks_he, data_dictionary_newsmap_he, levels = 3)
+    label_dfm_he <- dfm(label_toks_he)
+    feat_dfm_he <- dfm(toks_he, tolower = FALSE)
+    expect_equal(
+        as.character(predict(textmodel_newsmap(feat_dfm_he, label_dfm_he))),
+        "ie"
+    )
+})
+
+test_that("Arabic dictionary and prediction work correctly", {
+
+    skip_on_travis()
+    txt_ar <- c("هذا مقال عن أيرلندا.")
+    toks_ar <- tokens(txt_ar)
+    label_toks_ar <- tokens_lookup(toks_ar, data_dictionary_newsmap_ar, levels = 3)
+    label_dfm_ar <- dfm(label_toks_ar)
+    feat_dfm_ar <- dfm(toks_ar, tolower = FALSE)
+    expect_equal(
+        as.character(predict(textmodel_newsmap(feat_dfm_ar, label_dfm_ar))),
+        "ie"
+    )
+})
+
+test_that("Japanese dictionary and prediction work correctly", {
+
+    skip_on_travis()
+    txt_ja <- c("アイルランドに関するテキスト。")
+    toks_ja <- tokens(txt_ja)
     label_toks_ja <- tokens_lookup(toks_ja, data_dictionary_newsmap_ja, levels = 3)
     label_dfm_ja <- dfm(label_toks_ja)
 
@@ -50,8 +97,39 @@ test_that("test Japanese dictionary and prediction work correctly", {
     )
 })
 
-test_that("test methods on textmodel_newsmap works correctly", {
+test_that("Traditional Chinese dictionary and prediction work correctly", {
 
+    skip_on_travis()
+    txt_zh_tw <- c("這篇文章關於愛爾蘭。")
+    toks_zh_tw <- tokens(txt_zh_tw)
+    label_toks_zh_tw <- tokens_lookup(toks_zh_tw, data_dictionary_newsmap_zh_tw, levels = 3)
+    label_dfm_zh_tw <- dfm(label_toks_zh_tw)
+    feat_dfm_zh_tw <- dfm(toks_zh_tw, tolower = FALSE)
+
+    expect_equal(
+        as.character(predict(textmodel_newsmap(feat_dfm_zh_tw, label_dfm_zh_tw))),
+        "ie"
+    )
+})
+
+
+test_that("Simplified Chinese dictionary and prediction work correctly", {
+
+    skip_on_travis()
+    txt_zh_cn <- c("这篇文章关於爱尔兰。")
+    toks_zh_cn <- tokens(txt_zh_cn)
+    label_toks_zh_cn <- tokens_lookup(toks_zh_cn, data_dictionary_newsmap_zh_cn, levels = 3)
+    label_dfm_zh_cn <- dfm(label_toks_zh_cn)
+    feat_dfm_zh_cn <- dfm(toks_zh_cn, tolower = FALSE)
+
+    expect_equal(
+        as.character(predict(textmodel_newsmap(feat_dfm_zh_cn, label_dfm_zh_cn))),
+        "ie"
+    )
+})
+
+
+test_that("methods for textmodel_newsmap works correctly", {
     text <- c("Ireland is famous for Guinness.",
               "Guinness began retailing in India in 2007.",
               "Cork is an Irish coastal city.",
@@ -87,10 +165,42 @@ test_that("test methods on textmodel_newsmap works correctly", {
 
 })
 
-test_that("test raise error if dfm is empty", {
+test_that("textmodel_newsmap() raises error if dfm is empty", {
     expect_error(textmodel_newsmap(dfm_trim(dfm("a b c"), min_termfreq = 10), dfm("A")),
                  "x must have at least one non-zero feature")
 
     expect_error(textmodel_newsmap(dfm("a b c"), dfm_trim(dfm("A"), min_termfreq = 10)),
                  "y must have at least one non-zero feature")
 })
+
+test_that("predict() returns NA for documents without registered features", {
+
+    text <- c("Ireland is famous for Guinness.",
+              "Guinness began retailing in India in 2007.",
+              "Cork is an Irish coastal city.",
+              "Titanic departed Cork Harbour in 1912.")
+
+    toks <- tokens(text)
+    label_toks <- tokens_lookup(toks, data_dictionary_newsmap_en, levels = 3)
+    label_dfm <- dfm(label_toks)
+
+    dfmt_feat <- dfm(c("aa bb cc", "aa bb", "bb cc"))
+    dfmt_label <- dfm(c("A", "B", "B"), tolower = FALSE)
+    dfmt_new <- dfm(c("aa bb cc", "aa bb", "zz"))
+    map <- textmodel_newsmap(dfmt_feat, dfmt_label)
+    expect_equal(predict(map),
+                 c(text1 = "A", text2 = "B", text3 = "B"))
+    expect_equal(predict(map, newdata = dfmt_new),
+                 c(text1 = "A", text2 = "B", text3 = NA))
+    pred <- predict(map, confidence.fit = TRUE, newdata = dfmt_new)
+    expect_equal(pred$class,
+                 c(text1 = "A", text2 = "B", text3 = NA))
+    expect_equal(pred$confidence.fit,
+                 c(0.018, 0.048, NA), tolerance = 0.01)
+    expect_equal(predict(map, newdata = dfmt_new, rank = 2),
+                 c(text1 = "B", text2 = "A", text3 = NA))
+    expect_equal(as.numeric(predict(map, newdata = dfmt_new, type = "all")),
+                 c(0.018, -0.048, NA, -0.018, 0.048, NA), tolerance = 0.01)
+
+})
+